Provide a generic way of writing checkers
The parser is now splitted in two files, the common one and the
user-defined one which has to define a %public judgment rule.

The generated Ast.ml now provide types and functions to handle parsing
context and free/bound variables for a given checker.

The code is restructured and two Makefiles are provided, one in used to
build a checker if it is called inside the directory of a checker. The other
one can build the game maker and all the checkers by calling first the
game maker.